(PC-23806)[API] refactor: move titelive music search connector
