+++ api/src/pcapi/connectors/titelive.py
@@ -1,37 +1,21 @@
-import abc
 import datetime
 import enum
 import html
 import logging
 import typing
 
-import pydantic.v1 as pydantic_v1
 from urllib3 import exceptions as urllib3_exceptions
 
-from pcapi import repository
 from pcapi import settings
 from pcapi.core import logging as core_logging
 from pcapi.core.categories import subcategories
 from pcapi.core.offers import exceptions as offers_exceptions
 import pcapi.core.offers.models as offers_models
-import pcapi.core.providers.constants as providers_constants
-import pcapi.core.providers.models as providers_models
-import pcapi.core.providers.repository as providers_repository
-from pcapi.domain.music_types import MUSIC_SUB_TYPES_BY_SLUG
-from pcapi.domain.music_types import MUSIC_TYPES_BY_SLUG
-from pcapi.domain.music_types import OTHER_SHOW_TYPE_SLUG
 from pcapi.domain.titelive import read_things_date
-from pcapi.models import db
 from pcapi.utils import date as date_utils
 from pcapi.utils import requests
 from pcapi.utils.cache import get_from_cache
 from pcapi.utils.csr import get_closest_csr
-
-from .serialization.titelive_serializers import GenreTitelive
-from .serialization.titelive_serializers import TiteliveMusicArticle
-from .serialization.titelive_serializers import TiteliveMusicOeuvre
-from .serialization.titelive_serializers import TiteliveProductSearchResponse
-from .serialization.titelive_serializers import TiteliveSearchResultType
 
 
 logger = logging.getLogger(__name__)
@@ -203,237 +187,49 @@ class TiteliveBase(enum.Enum):
     MUSIC = "music"
 
 
-class TiteliveSearch(abc.ABC, typing.Generic[TiteliveSearchResultType]):
-    max_results_per_page = 25
-    titelive_base: TiteliveBase
+MAX_RESULTS_PER_PAGE = 25
 
-    def __init__(self) -> None:
-        self.provider = providers_repository.get_provider_by_name(providers_constants.TITELIVE_EPAGINE_PROVIDER_NAME)
 
-    def synchronize_products(self, from_date: datetime.date | None = None) -> None:
-        if from_date is None:
-            from_date = self.get_last_sync_date()
-
-        start_sync_event = self.log_sync_status(providers_models.LocalProviderEventType.SyncStart)
-        db.session.add(start_sync_event)
-
-        products_to_update_pages = self.get_updated_titelive_pages(from_date)
-        for titelive_page in products_to_update_pages:
-            with repository.transaction():
-                updated_products = self.upsert_titelive_page(titelive_page)
-                db.session.add_all(updated_products)
-
-        stop_sync_event = self.log_sync_status(providers_models.LocalProviderEventType.SyncEnd)
-        db.session.add(stop_sync_event)
-
-    def get_last_sync_date(self) -> datetime.date:
-        last_sync_event = (
-            providers_models.LocalProviderEvent.query.filter(
-                providers_models.LocalProviderEvent.provider == self.provider,
-                providers_models.LocalProviderEvent.type == providers_models.LocalProviderEventType.SyncEnd,
-                providers_models.LocalProviderEvent.payload == self.titelive_base.value,
-            )
-            .order_by(providers_models.LocalProviderEvent.id.desc())
-            .first()
+def search_products(titelive_base: TiteliveBase, from_date: datetime.date, page_index: int) -> dict[str, typing.Any]:
+    try:
+        url = f"{settings.TITELIVE_EPAGINE_API_URL}/search"
+        headers = {"Content-Type": "application/json", "Authorization": "Bearer {}".format(get_jwt_token())}
+        response = requests.get(
+            url,
+            headers=headers,
+            params={
+                "base": titelive_base.value,
+                "nombre": MAX_RESULTS_PER_PAGE,
+                "page": page_index,
+                "tri": "datemodification",
+                "tri_ordre": "asc",
+                "dateminm": date_utils.format_date_to_french_locale(from_date),
+            },
         )
-        if last_sync_event is None:
-            raise TiteliveDatabaseNotInitializedException()
-        return last_sync_event.date.date()
-
-    def log_sync_status(
-        self, provider_event_type: providers_models.LocalProviderEventType
-    ) -> providers_models.LocalProviderEvent:
-        return providers_models.LocalProviderEvent(
-            date=datetime.datetime.utcnow(),
-            payload=self.titelive_base.value,
-            provider=self.provider,
-            type=provider_event_type,
+    except (urllib3_exceptions.HTTPError, requests.exceptions.RequestException) as e:
+        logger.error(
+            "Titelive search: Network error",
+            extra={
+                "exception": e,
+                "alert": "Titelive error",
+                "error_type": "network",
+                "request_type": "search",
+            },
         )
+        raise requests.ExternalAPIException(is_retryable=True) from e
 
-    def get_updated_titelive_pages(
-        self,
-        from_date: datetime.date,
-    ) -> typing.Iterator[TiteliveProductSearchResponse[TiteliveSearchResultType]]:
-        page_index = 1
-        titelive_product_page = None
-        while titelive_product_page is None or len(titelive_product_page.result) == self.max_results_per_page:
-            titelive_json_response = self.get_titelive_search_json(from_date, page_index)
-            titelive_product_page = self.deserialize_titelive_search_json(titelive_json_response)
-            allowed_product_page = self.filter_allowed_products(titelive_product_page)
-            yield allowed_product_page
-            page_index += 1
-
-    def get_titelive_search_json(self, from_date: datetime.date, page_index: int) -> dict[str, typing.Any]:
-        try:
-            url = f"{settings.TITELIVE_EPAGINE_API_URL}/search"
-            headers = {"Content-Type": "application/json", "Authorization": "Bearer {}".format(get_jwt_token())}
-            response = requests.get(
-                url,
-                headers=headers,
-                params={
-                    "base": self.titelive_base.value,
-                    "nombre": self.max_results_per_page,
-                    "page": page_index,
-                    "tri": "datemodification",
-                    "tri_ordre": "asc",
-                    "dateminm": date_utils.format_date_to_french_locale(from_date),
-                },
-            )
-        except (urllib3_exceptions.HTTPError, requests.exceptions.RequestException) as e:
-            logger.error(
-                "Titelive search: Network error",
-                extra={
-                    "exception": e,
-                    "alert": "Titelive error",
-                    "error_type": "network",
-                    "request_type": "search",
-                },
-            )
-            raise requests.ExternalAPIException(is_retryable=True) from e
-
-        if not response.ok:
-            logger.error(
-                "Titelive search: External error %s",
-                response.status_code,
-                extra={
-                    "alert": "Titelive error",
-                    "error_type": "http",
-                    "status_code": response.status_code,
-                    "request_type": "search",
-                    "response_text": response.text,
-                },
-            )
-            raise requests.ExternalAPIException(True, {"status_code": response.status_code})
-
-        return response.json()
-
-    @abc.abstractmethod
-    def deserialize_titelive_search_json(
-        self, titelive_json_response: dict[str, typing.Any]
-    ) -> TiteliveProductSearchResponse[TiteliveSearchResultType]:
-        raise NotImplementedError()
-
-    def filter_allowed_products(
-        self,
-        titelive_product_page: TiteliveProductSearchResponse[TiteliveSearchResultType],
-    ) -> TiteliveProductSearchResponse[TiteliveSearchResultType]:
-        return titelive_product_page
-
-    def upsert_titelive_page(
-        self,
-        titelive_page: TiteliveProductSearchResponse[TiteliveSearchResultType],
-    ) -> list[offers_models.Product]:
-        titelive_results = titelive_page.result
-        titelive_eans = [article.gencod for result in titelive_results for article in result.article]
-
-        products = offers_models.Product.query.filter(
-            offers_models.Product.extraData["ean"].astext.in_(titelive_eans),
-            offers_models.Product.lastProvider == self.provider,  # pylint: disable=comparison-with-callable
-        ).all()
-        products_by_ean: dict[str, offers_models.Product] = {p.extraData["ean"]: p for p in products}
-        for titelive_search_result in titelive_results:
-            products_by_ean = self.upsert_titelive_result_in_dict(titelive_search_result, products_by_ean)
-
-        return list(products_by_ean.values())
-
-    @abc.abstractmethod
-    def upsert_titelive_result_in_dict(
-        self, titelive_search_result: TiteliveSearchResultType, products_by_ean: dict[str, offers_models.Product]
-    ) -> dict[str, offers_models.Product]:
-        raise NotImplementedError()
-
-
-class TiteliveMusicSearch(TiteliveSearch[TiteliveMusicOeuvre]):
-    titelive_base = TiteliveBase.MUSIC
-
-    def deserialize_titelive_search_json(
-        self, titelive_json_response: dict[str, typing.Any]
-    ) -> TiteliveProductSearchResponse[TiteliveMusicOeuvre]:
-        return pydantic_v1.parse_obj_as(TiteliveProductSearchResponse[TiteliveMusicOeuvre], titelive_json_response)
-
-    def upsert_titelive_result_in_dict(
-        self, titelive_search_result: TiteliveMusicOeuvre, products_by_ean: dict[str, offers_models.Product]
-    ) -> dict[str, offers_models.Product]:
-        title = titelive_search_result.titre
-        genre_titelive = self.get_genre_titelive(titelive_search_result)
-        for article in titelive_search_result.article:
-            ean = article.gencod
-            product = products_by_ean.get(ean)
-            if product is None:
-                products_by_ean[ean] = self.create_product(article, title, genre_titelive)
-            else:
-                products_by_ean[ean] = self.update_product(article, title, genre_titelive, product)
-
-        return products_by_ean
-
-    def get_genre_titelive(self, titelive_search_result: TiteliveMusicOeuvre) -> GenreTitelive | None:
-        titelive_gtl = next(
-            (article.gtl for article in titelive_search_result.article if article.gtl is not None), None
+    if not response.ok:
+        logger.error(
+            "Titelive search: External error %s",
+            response.status_code,
+            extra={
+                "alert": "Titelive error",
+                "error_type": "http",
+                "status_code": response.status_code,
+                "request_type": "search",
+                "response_text": response.text,
+            },
         )
-        if not titelive_gtl:
-            logger.warning("Genre Titelive not found for music %s", titelive_search_result.titre)
-            return None
+        raise requests.ExternalAPIException(True, {"status_code": response.status_code})
 
-        most_precise_genre = max(titelive_gtl.first.values(), key=lambda g: g.code)
-        return most_precise_genre
-
-    def create_product(
-        self, article: TiteliveMusicArticle, title: str, genre: GenreTitelive | None
-    ) -> offers_models.Product:
-        return offers_models.Product(
-            description=article.resume,
-            extraData=self.build_music_extra_data(article, genre),
-            idAtProviders=article.gencod,
-            lastProvider=self.provider,
-            name=title,
-            subcategoryId=self.parse_titelive_product_format(article.codesupport),
-        )
-
-    def update_product(
-        self, article: TiteliveMusicArticle, title: str, genre: GenreTitelive | None, product: offers_models.Product
-    ) -> offers_models.Product:
-        product.description = article.resume
-        if product.extraData is None:
-            product.extraData = offers_models.OfferExtraData()
-        product.extraData.update(self.build_music_extra_data(article, genre))
-        product.idAtProviders = article.gencod
-        product.name = title
-        product.subcategoryId = self.parse_titelive_product_format(article.codesupport)
-
-        return product
-
-    def build_music_extra_data(
-        self, article: TiteliveMusicArticle, genre: GenreTitelive | None
-    ) -> offers_models.OfferExtraData:
-        gtl_id = genre.code.zfill(8) if genre else None
-        music_type_id, music_subtype_id = self.parse_titelive_music_genre(genre)
-
-        return offers_models.OfferExtraData(
-            artist=article.artiste,
-            author=article.compositeur,
-            comment=article.commentaire,
-            date_parution=article.dateparution.isoformat() if article.dateparution else None,
-            disponibility=article.libelledispo,
-            distributeur=article.distributeur,
-            ean=article.gencod,
-            editeur=article.editeur,
-            gtl_id=gtl_id,
-            music_label=article.label,
-            musicSubType=str(music_subtype_id),
-            musicType=str(music_type_id),
-            nb_galettes=article.nb_galettes,
-            performer=article.interprete,
-        )
-
-    def parse_titelive_music_genre(self, genre: GenreTitelive | None) -> tuple[int, int]:
-        return (
-            MUSIC_TYPES_BY_SLUG[OTHER_SHOW_TYPE_SLUG].code,
-            MUSIC_SUB_TYPES_BY_SLUG[OTHER_SHOW_TYPE_SLUG].code,
-        )
-
-    def parse_titelive_product_format(self, codesupport: str) -> str:
-        return subcategories.SUPPORT_PHYSIQUE_MUSIQUE.id
-
-
-class TiteliveDatabaseNotInitializedException(Exception):
-    pass
+    return response.json()
